requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-yasg==1.21.7
django-filter==23.5
Faker==20.1.0
psycopg2-binary==2.9.9 